Fix Static Analysis issues on Coverity#4
:Release Notes:
Fix static analysis issues

:Detailed Notes:
CID : [9156256], [9156253], [9156242], [9156228]

:Testing Performed:
Local test passed

:QA Notes:
N/A

:Issues Addressed:
[WRP-31289] Fix coverity issues

+++ src/util/NativeProcess.cpp
@@ -99,7 +99,7 @@ bool NativeProcess::run()
         0,
     };
     string params = "";
-    int index = 0;
+    long long index = 0;
 
     GError* gerr = NULL;
     argv[0] = m_command.c_str();
